requirements.txt
googletrans
pyowm
python-dotenv
discord.py
pycountry
discord.py[voice]
youtube_dl
spotipy